+++ src/main.ts
@@ -15,24 +15,30 @@ export default class KoboHighlightsImporter extends Plugin {
 		addIcon("e-reader", EREADER_ICON_PATH);
 		await this.loadSettings();
 
-		// This creates an icon in the left ribbon.
+		const saveSettings = () => this.saveSettings();
+
 		const KoboHighlightsImporterIconEl = this.addRibbonIcon(
 			"e-reader",
 			"Import from Kobo",
 			() => {
-				// Called when the user clicks the icon.
-				new ExtractHighlightsModal(this.app, this.settings).open();
+				new ExtractHighlightsModal(
+					this.app,
+					this.settings,
+					saveSettings,
+				).open();
 			},
 		);
-		// Perform additional things with the ribbon
 		KoboHighlightsImporterIconEl.addClass("kobo-highlights-importer-icon");
 
-		// This adds a simple command that can be triggered anywhere
 		this.addCommand({
 			id: "import-from-kobo-sqlite",
 			name: "Import from Kobo",
 			callback: () => {
-				new ExtractHighlightsModal(this.app, this.settings).open();
+				new ExtractHighlightsModal(
+					this.app,
+					this.settings,
+					saveSettings,
+				).open();
 			},
 		});